+++ changelog.md
@@ -1,6 +1,11 @@
 ![Mollie](https://www.mollie.nl/files/Mollie-Logo-Style-Small.png)
 
 # Changelog #
+## Changes in release 5.2.0 ##
++ Fixed issue where customer id was not added when using saved credit card for order API.
++ Added QR code option for bancontact payment on live mode
++ Added in3 payment method restriction for using only order API.
+
 ## Changes in release 5.1.0 ##
 + Added order refund statuses for methods that use Order API.
 + Fixed refund and cancel information text in order page.
@@ -9,6 +14,8 @@
 + Fixed issue where order API would not update status if refunded.
 + Fixed issue where credit card token was not added if user is not saved.
 + No longer try to create second mollie transaction without number if first failed and if it has credit card.
++ Ignore beta releases when showing warning about new version in module.
++ Added fix for custom checkout modules and apple pay payment.
 
 ## Changes in release 5.0.1 ##
 + Fixed issue where order API refund would fail on webhook call.